calendai_backend/requirements.txt
fastapi
fastapi[standard]
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth
google-auth-oauthlib
google-api-core
google-api-python-client
dotenv-python